(#57) Introduce scene module
